radiologytool/requirements.txt
flask==2.0.1
openai==1.12.0
python-dotenv==0.19.0
gunicorn==20.1.0
werkzeug==2.0.3 